# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ItineraryResponse": {
        "properties": {
          "itinerary_revision": {
            "title": "Itinerary Revision",
            "type": "integer"
          },
          "legs": {
            "items": {
              "$ref": "#/components/schemas/ItineraryTravelLegResponse"
            },
            "title": "Legs",
            "type": "array"
          },
          "stops": {
            "items": {
              "$ref": "#/components/schemas/ItineraryStopResponse"
            },
            "title": "Stops",
            "type": "array"
          },
          "trip_id": {
            "format": "uuid",
            "title": "Trip Id",
            "type": "string"
          }
        },
        "required": [
          "trip_id",
          "itinerary_revision",
          "stops",
          "legs"
        ],
        "title": "ItineraryResponse",
        "type": "object"
      },
...
      "ItineraryStopCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "incoming_travel": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ItineraryTravelReplaceRequest"
              },
              {
                "type": "null"
              }
            ]
          },
          "location": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LocationPlaceInput"
              },
              {
                "$ref": "#/components/schemas/LocationCoordinatesInput"
              }
            ],
            "title": "Location"
          },
          "notes": {
            "default": "",
            "title": "Notes",
            "type": "string"
          },
          "outgoing_travel": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ItineraryTravelReplaceRequest"
              },
              {
                "type": "null"
              }
            ]
          },
          "placement": {
            "$ref": "#/components/schemas/ItineraryPlacement"
          },
          "planned_nights": {
            "minimum": 0.0,
            "title": "Planned Nights",
            "type": "integer"
          },
          "title": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "location",
          "title",
          "planned_nights",
          "placement"
        ],
        "title": "ItineraryStopCreateRequest",
        "type": "object"
      },
      "ItineraryStopDetailResponse": {
        "properties": {
          "incoming_leg": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ItineraryTravelLegResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "outgoing_leg": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ItineraryTravelLegResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "stop": {
            "$ref": "#/components/schemas/ItineraryStopResponse"
          }
        },
        "required": [
          "stop",
          "incoming_leg",
          "outgoing_leg"
        ],
        "title": "ItineraryStopDetailResponse",
        "type": "object"
      },
      "ItineraryStopResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "$ref": "#/components/schemas/UserSummaryResponse"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "location": {
            "$ref": "#/components/schemas/LocationResponse"
          },
          "notes": {
            "title": "Notes",
            "type": "string"
          },
          "planned_nights": {
            "title": "Planned Nights",
            "type": "integer"
          },
          "planned_start_date": {
            "format": "date",
            "title": "Planned Start Date",
            "type": "string"
          },
          "same_day_position": {
            "title": "Same Day Position",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "trip_id": {
            "format": "uuid",
            "title": "Trip Id",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "visited": {
            "title": "Visited",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "trip_id",
          "same_day_position",
          "location",
          "title",
          "notes",
          "planned_start_date",
          "planned_nights",
          "visited",
          "created_by",
          "created_at",
          "updated_at"
        ],
        "title": "ItineraryStopResponse",
        "type": "object"
      },
      "ItineraryStopUpdateRequest": {
        "additionalProperties": false,
        "properties": {
          "incoming_travel": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ItineraryTravelReplaceRequest"
              },
              {
                "type": "null"
              }
            ]
          },
          "location": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LocationPlaceInput"
              },
              {
                "$ref": "#/components/schemas/LocationCoordinatesInput"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "outgoing_travel": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ItineraryTravelReplaceRequest"
              },
              {
                "type": "null"
              }
            ]
          },
          "placement": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ItineraryPlacement"
              },
              {
                "type": "null"
              }
            ]
          },
          "planned_nights": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Planned Nights"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "ItineraryStopUpdateRequest",
        "type": "object"
      },
      "ItineraryTravelLegResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "from_stop_id": {
            "format": "uuid",
            "title": "From Stop Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "notes": {
            "title": "Notes",
            "type": "string"
          },
          "operator": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operator"
          },
          "reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference"
          },
          "route": {
            "$ref": "#/components/schemas/ItineraryTravelRouteResponse"
          },
          "to_stop_id": {
            "format": "uuid",
            "title": "To Stop Id",
            "type": "string"
          },
          "travel_mode": {
            "$ref": "#/components/schemas/TravelMode"
          },
          "trip_id": {
            "format": "uuid",
            "title": "Trip Id",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "trip_id",
          "from_stop_id",
          "to_stop_id",
          "travel_mode",
          "notes",
          "operator",
          "reference",
          "route",
          "created_at",
          "updated_at"
        ],
        "title": "ItineraryTravelLegResponse",
        "type": "object"
      },
      "ItineraryTravelReplaceRequest": {
        "additionalProperties": false,
        "properties": {
          "notes": {
            "default": "",
            "title": "Notes",
            "type": "string"
          },
          "operator": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operator"
          },
          "reference": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference"
          },
          "travel_mode": {
            "$ref": "#/components/schemas/TravelMode"
          }
        },
        "required": [
          "travel_mode"
        ],
        "title": "ItineraryTravelReplaceRequest",
        "type": "object"
      },
...
      "TravelMode": {
        "description": "Transport mode shared by itinerary legs, post routes, and GPS samples.\n\nValues are stored as plain strings in ``String(32)`` columns, so adding a\nmember is a code change only and never needs a migration.",
        "enum": [
          "UNKNOWN",
          "WALK",
          "BIKE",
          "MOTORCYCLE",
          "CAR",
          "BUS",
          "TRAIN",
          "FERRY",
          "FLIGHT",
          "OTHER"
        ],
        "title": "TravelMode",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/login/access-token"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "OpenVoyage API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/trips/{trip_id}/itinerary": {
      "get": {
        "operationId": "get_itinerary_api_v1_trips__trip_id__itinerary_get",
        "parameters": [
          {
            "in": "path",
            "name": "trip_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Trip Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Trip-Share-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Trip-Share-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItineraryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Itinerary",
        "tags": [
          "itinerary"
        ]
      }
    },
    "/api/v1/trips/{trip_id}/itinerary/legs/{leg_id}": {
      "get": {
        "operationId": "get_leg_api_v1_trips__trip_id__itinerary_legs__leg_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "trip_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Trip Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "leg_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Leg Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Trip-Share-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Trip-Share-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItineraryTravelLegResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Leg",
        "tags": [
          "itinerary"
        ]
      },
      "put": {
        "operationId": "replace_leg_api_v1_trips__trip_id__itinerary_legs__leg_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "trip_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Trip Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "leg_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Leg Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItineraryTravelReplaceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItineraryTravelLegResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Replace Leg",
        "tags": [
          "itinerary"
        ]
      }
    },
    "/api/v1/trips/{trip_id}/itinerary/legs/{leg_id}/route-refresh": {
      "post": {
        "operationId": "refresh_leg_route_api_v1_trips__trip_id__itinerary_legs__leg_id__route_refresh_post",
        "parameters": [
          {
            "in": "path",
            "name": "trip_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Trip Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "leg_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Leg Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItineraryTravelLegResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Refresh Leg Route",
        "tags": [
          "itinerary"
        ]
      }
    },
    "/api/v1/trips/{trip_id}/itinerary/stops": {
      "post": {
        "operationId": "create_stop_api_v1_trips__trip_id__itinerary_stops_post",
        "parameters": [
          {
            "in": "path",
            "name": "trip_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Trip Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItineraryStopCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItineraryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create Stop",
        "tags": [
          "itinerary"
        ]
      }
    },
    "/api/v1/trips/{trip_id}/itinerary/stops/{stop_id}": {
      "delete": {
        "operationId": "delete_stop_api_v1_trips__trip_id__itinerary_stops__stop_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "trip_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Trip Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "stop_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Stop Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItineraryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Delete Stop",
        "tags": [
          "itinerary"
        ]
      },
      "get": {
        "operationId": "get_stop_api_v1_trips__trip_id__itinerary_stops__stop_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "trip_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Trip Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "stop_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Stop Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Trip-Share-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Trip-Share-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItineraryStopDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Stop",
        "tags": [
          "itinerary"
        ]
      },
      "patch": {
        "operationId": "update_stop_api_v1_trips__trip_id__itinerary_stops__stop_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "trip_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Trip Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "stop_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Stop Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItineraryStopUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItineraryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update Stop",
        "tags": [
          "itinerary"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:8000"
    }
  ]
}